Add isSetTo with parent parameter to CSHasChangeValue
This commit introduces a new overload for the `isSetTo` function in the `CSHasChangeValue` class. The new function takes a `parent` parameter of type `CSHasRegistrations` in addition to the value to check against. This allows for creating `CSHasChangeValue<Boolean>` instances that track whether a value is set to a specific value within the context of a parent registration.