Refactor: Introduce event for added items in list factory
Adds an `eventAdded` parameter to the `factory` and `update` methods for `MutableList<T : CSHasDestruct>`. This event is triggered when a new item is added to the list, providing the newly added item as an argument.

This change allows for more reactive handling of list modifications in `PadController` and `PadControllerRow` by introducing `eventRowAdded` and `eventKeyAdded` respectively.