Add `eventRemoved` parameter to `MutableList.update` and `factory`
This commit introduces an `eventRemoved` parameter to the `update` and `factory` extension functions for `MutableList<T : CSHasDestruct>`.

This allows for custom actions to be performed when an item is removed from the list, in addition to calling `destruct()` on the removed item. The `eventRemoved` is an optional `CSEvent<Int>` in the `factory` function, defaulting to an empty event.