Make `CSEventListener` an operator function and add `fireOnMain`
This commit changes `CSEventListener.invoke` to an operator function for more idiomatic Kotlin usage.

A new function `fireOnMain` has been added to `CSEventImpl`. This function allows event firing to be scheduled on the main thread if an associated `CSHasDestruct` parent is provided.

`CSSafePropertyImpl` has been updated to utilize `fireOnMain` for its `eventChange`, ensuring that property change events are dispatched on the main thread when a parent is available.

The companion object factory methods in `CSEvent` have been simplified.