Refactor: Use `out` variance for `CSHasChangeValue` type parameter
This commit updates the `destructPrevious` extension functions in `CSHasChangeValue+.kt` to use `out` variance for the type parameter `T` in `CSHasChangeValue`. This allows for greater flexibility when using these functions with subtypes.

A new extension function `delegateDestruct` has been added to `CSHasChangeValue.kt`. This function combines the functionality of `delegate` and `destructPrevious`, providing a more concise way to delegate a `CSHasChangeValue` and ensure the previous value is destructed.